SmartCity/requirements.txt
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
ConfigArgParse==1.5.3
contourpy==1.0.5
cycler==0.11.0
dbus-python==1.2.16
efficientnet-pytorch==0.7.1
flake8==5.0.4
fonttools==4.37.3
idna==3.4
imutils==0.5.4
iniconfig==1.1.1
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.6.0
mccabe==0.7.0
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
py-cpuinfo==8.0.0
pycairo==1.21.0
pycocotools==2.0.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyGObject==3.36.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytest==7.1.3
pytest-benchmark==3.4.1
python-dateutil==2.8.2
requests==2.28.1
scipy==1.9.1
six==1.16.0
tabulate==0.8.10
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
gstreamer-python @ git+https://github.com/jackersson/gstreamer-python.git#egg=gstreamer-python